Adds Jest testing framework
Adds Jest as the testing framework for the project.

Includes configuration and scripts for running tests, watching tests, and generating coverage reports.

Also updates the LicenseAgent class to be testable.
Specifically, it makes the config property public to allow for easier testing.

Removes debug and warn logs that pollutes output and are not useful for users of the library.